day15/test_aoc_202315.py
import pathlib

import aoc_202315 as aoc
import pytest

PUZZLE_DIR = pathlib.Path(__file__).parent


@pytest.fixture
def example1():
	puzzle_input = (PUZZLE_DIR / "example1.txt").read_text().strip()
	return aoc.parse(puzzle_input)


@pytest.fixture
def example2():
	puzzle_input = (PUZZLE_DIR / "input1.txt").read_text().strip()
	return aoc.parse(puzzle_input)


# @pytest.mark.skip(reason="Not implemented")
def test_parse_example1(example1):
	"""Test that input is parsed properly"""
	assert len(example1) == 11
	assert example1[0] == "rn=1"
	assert example1[-1] == "ot=7"


def test_hash1():
	inp = "HASH"
	actual = aoc.my_hash(inp)
	assert actual == 52


def test_parse_mirror():
	assert aoc.parse_mirror("rn=1") == ("rn", 1)
	assert aoc.parse_mirror("cm-") == ("cm", None)


# @pytest.mark.skip(reason="Not implemented")
def test_part1_example1(example1):
	"""Test part 1 on example input"""
	assert aoc.part1(example1) == 1320


# @pytest.mark.skip(reason="Not implemented")
def test_part2_example1(example1):
	"""Test part 2 on example input"""
	assert aoc.part2(example1) == 145


# @pytest.mark.skip(reason="Not implemented")
def test_part2_example2(example2):
	"""Test part 2 on example input"""
	assert aoc.part2(example2) == 145

